Update EvaluationTemplate to have clearer links to each UseCase
The idea is to use the evaluation template to fill out for each project
codebase evaluation we do, but have an easy link back to the use case it
refers to.

As I was making this edit, I added a few FIXMEs to things that need
clarification and/or completion.